+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+## next
+
+- Fixed crash when `localStorage` is not available (e.g. due CSP restrictions)
+
 ## 1.0.0-beta.44 (21-10-2020)
 
 - Added option to control localStorage usage for darkmode state