+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## next
+## 1.0.0-beta.45 (21-10-2020)
 
 - Fixed crash when `localStorage` is not available (e.g. due CSP restrictions)
 - Fixed `darkmode` and `darkmodePersistent` settings for `App` to use defaults when coresponding values in options are `undefined`